+++ nidup/pysc2/smart_orders.py
@@ -1,7 +1,7 @@
 
 import random
 from pysc2.lib import actions
-from nidup.pysc2.actions import TerranActions, TerranActionIds
+from nidup.pysc2.wrapper.actions import TerranActions, TerranActionIds
 from nidup.pysc2.wrapper.observations import Observations
 from nidup.pysc2.wrapper.unit_types import UnitTypeIds
